[PWGMM][dndeta-mft] MegaLinter correction

+++ PWGMM/Mult/Tasks/dndeta-mft.cxx
@@ -10,6 +10,10 @@
 // or submit itself to any jurisdiction.
 
 #include <cmath>
+// for CCDB access
+#include <chrono>
+#include "CCDB/BasicCCDBManager.h"
+
 #include "Framework/Configurable.h"
 #include "Framework/AnalysisTask.h"
 #include "Framework/AnalysisDataModel.h"
@@ -25,9 +29,7 @@
 #include "CommonConstants/MathConstants.h"
 #include "TDatabasePDG.h"
 #include "MathUtils/Utils.h"
-// for CCDB access
-#include <CCDB/BasicCCDBManager.h>
-#include <chrono>
+
 
 #include "bestCollisionTable.h"
 
@@ -56,11 +58,11 @@ struct PseudorapidityDensityMFT {
   Service<ccdb::BasicCCDBManager> ccdb;
   Configurable<std::string> path{"ccdb-path", "Users/s/sherrman/My/Object", "base path to the ccdb object"};
   Configurable<std::string> url{"ccdb-url", "http://alice-ccdb.cern.ch", "url of the ccdb repository"};
-  Configurable<long> nolaterthan{"ccdb-no-later-than", std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count(), "latest acceptable timestamp of creation for the object"};
+  Configurable<int64> nolaterthan{"ccdb-no-later-than", std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count(), "latest acceptable timestamp of creation for the object"};
 
   // the histogram has been previously stored in the CCDB
   TH1D* histoReweight = nullptr;
-  std::vector<long long unsigned> ambTrackIds;
+  std::vector<uint64_t> ambTrackIds;
   int counter = 0;
   //------
 
@@ -332,7 +334,7 @@ struct PseudorapidityDensityMFT {
       auto p = pdg->GetParticle(particle.pdgCode());
       auto charge = 0;
       if (p != nullptr) {
-        charge = (int)p->Charge();
+        charge = static_cast<int>(p->Charge());
       }
       if (std::abs(charge) < 3.) {
         continue;
